Moved CustomWebApplicationFactory into IntegrationTests.Util
