+++ test/Altinn.Notifications.Tests/IntegrationTests/Utils/CustomWebApplicationFactory.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
 
-namespace Altinn.Notifications.Tests
+namespace Altinn.Notifications.Tests.IntegrationTests.Utils
 {
     public class CustomWebApplicationFactory<TStartup> : WebApplicationFactory<TStartup>
         where TStartup : class
